requirements.txt
# Core dependencies - Browser-use
browser-use>=0.1.41
python-dotenv>=1.1.0
pydantic>=2.10.6
PyYAML>=6.0.2

# Core dependencies - brightdata mcp


# Web scraping utilities
beautifulsoup4>=4.13.4
requests>=2.32.3

# LLM integrations
langchain>=0.3.22
langchain-openai>=0.3.11
langchain-google-genai>=2.1.2
langchain-anthropic>=0.3.3
langchain-ollama>=0.3.0

# Vector database
qdrant-client>=1.13.3

# for testing locally
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0


# For PDF functionality (optional)
# weasyprint>=60.1